+++ http/src/test/scala/com/twitter/finatra/http/EmbeddedHttpServer.scala
@@ -187,7 +187,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a GET request against the embedded server serializing the normalized
+    * Performs a GET request to the embedded server serializing the normalized
     * response#contentString into an instance of type [[ResponseType]].
     *
     * @see [[com.twitter.finatra.json.FinatraObjectMapper]]#parse[T: Manifest](string: String)
@@ -208,7 +208,7 @@ class EmbeddedHttpServer(
     * @tparam ResponseType - parse the response#contentString into type [[ResponseType]]
     * @return instance of type [[ResponseType]] serialized from the the response#contentString.
     */
-  def httpGetJson[ResponseType: Manifest](
+  def httpGetJson[ResponseType : Manifest](
     path: String,
     accept: MediaType = null,
     headers: Map[String, String] = Map(),
@@ -233,7 +233,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a POST request against the embedded server.
+    * Performs a POST request to the embedded server.
     *
     * @param path - URI of the request
     * @param postBody - body of the POST request
@@ -277,7 +277,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a POST request against the embedded server serializing the normalized
+    * Performs a POST request to the embedded server serializing the normalized
     * response#contentString into an instance of type [[ResponseType]].
     *
     * @see [[com.twitter.finatra.json.FinatraObjectMapper]]#parse[T: Manifest](string: String)
@@ -298,7 +298,7 @@ class EmbeddedHttpServer(
     * @tparam ResponseType - parse the response#contentString into type [[ResponseType]]
     * @return instance of type [[ResponseType]] serialized from the the response#contentString.
     */
-  def httpPostJson[ResponseType: Manifest](
+  def httpPostJson[ResponseType : Manifest](
     path: String,
     postBody: String,
     suppress: Boolean = false,
@@ -319,7 +319,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a PUT request against the embedded server.
+    * Performs a PUT request to the embedded server.
     *
     * @param path - URI of the request
     * @param putBody - the body of the PUT request
@@ -363,7 +363,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a PUT request against the embedded server serializing the normalized
+    * Performs a PUT request to the embedded server serializing the normalized
     * response#contentString into an instance of type [[ResponseType]].
     *
     * @see [[com.twitter.finatra.json.FinatraObjectMapper]]#parse[T: Manifest](string: String)
@@ -384,7 +384,7 @@ class EmbeddedHttpServer(
     * @tparam ResponseType - parse the response#contentString into type [[ResponseType]]
     * @return instance of type [[ResponseType]] serialized from the the response#contentString.
     */
-  def httpPutJson[ResponseType: Manifest](
+  def httpPutJson[ResponseType : Manifest](
     path: String,
     putBody: String,
     suppress: Boolean = false,
@@ -405,7 +405,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a DELETE request against the embedded server.
+    * Performs a DELETE request to the embedded server.
     *
     * @param path - URI of the request
     * @param deleteBody - the body of the DELETE request
@@ -462,7 +462,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a DELETE request against the embedded server serializing the normalized
+    * Performs a DELETE request to the embedded server serializing the normalized
     * response#contentString into an instance of type [[ResponseType]].
     *
     * @see [[com.twitter.finatra.json.FinatraObjectMapper]]#parse[T: Manifest](string: String)
@@ -483,7 +483,7 @@ class EmbeddedHttpServer(
     * @tparam ResponseType - parse the response#contentString into type [[ResponseType]]
     * @return instance of type [[ResponseType]] serialized from the the response#contentString.
     */
-  def httpDeleteJson[ResponseType: Manifest](
+  def httpDeleteJson[ResponseType : Manifest](
     path: String,
     deleteBody: String,
     suppress: Boolean = false,
@@ -504,7 +504,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a OPTIONS request against the embedded server.
+    * Performs a OPTIONS request to the embedded server.
     *
     * @param path - URI of the request
     * @param accept - add request Accept header with the given [[com.google.common.net.MediaType]]
@@ -540,7 +540,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a PATCH request against the embedded server.
+    * Performs a PATCH request to the embedded server.
     *
     * @param path - URI of the request
     * @param patchBody - the body of the PATCH request
@@ -584,7 +584,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a PATCH request against the embedded server serializing the normalized
+    * Performs a PATCH request to the embedded server serializing the normalized
     * response#contentString into an instance of type [[ResponseType]].
     *
     * @see [[com.twitter.finatra.json.FinatraObjectMapper]]#parse[T: Manifest](string: String)
@@ -605,7 +605,7 @@ class EmbeddedHttpServer(
     * @tparam ResponseType - parse the response#contentString into type [[ResponseType]]
     * @return instance of type [[ResponseType]] serialized from the the response#contentString.
     */
-  def httpPatchJson[ResponseType: Manifest](
+  def httpPatchJson[ResponseType : Manifest](
     path: String,
     patchBody: String,
     suppress: Boolean = false,
@@ -626,7 +626,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a HEAD request against the embedded server.
+    * Performs a HEAD request to the embedded server.
     *
     * @param path - URI of the request
     * @param accept - add request Accept header with the given [[com.google.common.net.MediaType]]
@@ -662,7 +662,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a form POST request against the embedded server.
+    * Performs a form POST request to the embedded server.
     *
     * @param path - URI of the request
     * @param params - a Map[String,String] of form params to send in the request
@@ -700,7 +700,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Performs a multi-part form POST request against the embedded server.
+    * Performs a multi-part form POST request to the embedded server.
     *
     * @param path - URI of the request
     * @param params - a Seq of [[com.twitter.finagle.http.FormElement]] to send in the request
@@ -736,7 +736,7 @@ class EmbeddedHttpServer(
   }
 
   /**
-    * Sends the given [[com.twitter.finagle.http.Request]] against the embedded server.
+    * Sends the given [[com.twitter.finagle.http.Request]] to the embedded server.
     *
     * @param request - built [[com.twitter.finagle.http.Request]] to send to the embedded server
     * @param suppress - suppress http client logging
@@ -771,6 +771,43 @@ class EmbeddedHttpServer(
     request: Request): Response = {
 
     httpRequest(request, suppress = false)
+  }
+
+  /**
+   * Sends the given [[com.twitter.finagle.http.Request]] to the embedded server
+   * serializing the normalized response#contentString into an instance of type [[ResponseType]]
+   *
+   * @see [[com.twitter.finatra.json.FinatraObjectMapper]]#parse[T: Manifest](string: String)
+   * @param request - built [[com.twitter.finagle.http.Request]] to send to the embedded server
+   * @param suppress - suppress http client logging
+   * @param andExpect - expected [[com.twitter.finagle.http.Status]] value
+   * @param withLocation - expected response Location header value
+   * @param withBody - expected body as a String
+   * @param withJsonBody - expected body as JSON
+   * @param withJsonBodyNormalizer - normalizer to use in conjunction with withJsonBody
+   * @param normalizeJsonParsedReturnValue - if the normalizer SHOULD be applied on the parsing of the
+   *                                       response#contentString into type [[ResponseType]], default = false
+   * @param withErrors - expected errors
+   * @param routeToAdminServer - force the request to the admin interface of the embedded server, false by default
+   * @param secure - use the https port to address the embedded server, default = None
+   * @return instance of type [[ResponseType]] serialized from the response#contentString
+   */
+  def httpRequestJson[ResponseType : Manifest](
+    request: Request,
+    suppress: Boolean = false,
+    andExpect: Status = null,
+    withLocation: String = null,
+    withBody: String = null,
+    withJsonBody: String = null,
+    withJsonBodyNormalizer: JsonNode => JsonNode = null,
+    normalizeJsonParsedReturnValue: Boolean = false,
+    withErrors: Seq[String] = null,
+    routeToAdminServer: Boolean = false,
+    secure: Option[Boolean] = None): ResponseType = {
+
+    assert(manifest[ResponseType] != manifest[Nothing], "httpRequestJson requires a type-param to parse the JSON response into, e.g. http<Method>Json[MyCaseClass] or http<Method>Json[JsonNode]")
+    val response = httpRequest(request, suppress, andExpect, withLocation, withBody, withJsonBody, withJsonBodyNormalizer, withErrors, routeToAdminServer, secure)
+    jsonParseWithNormalizer(response, withJsonBodyNormalizer, normalizeJsonParsedReturnValue)
   }
 
   /* Private */
@@ -863,7 +900,7 @@ class EmbeddedHttpServer(
       headers
   }
 
-  private def jsonParseWithNormalizer[T: Manifest](
+  private def jsonParseWithNormalizer[T : Manifest](
     response: Response,
     normalizer: JsonNode => JsonNode,
     normalizeParsedJsonNode: Boolean) = {